requirements.txt
certifi==2020.12.5
chardet==4.0.0
dnspython==2.1.0
idna==2.10
pkg-resources==0.0.0
pymongo==3.11.3
pyTelegramBotAPI==3.7.6
python-dotenv==0.15.0
requests==2.25.1
schedule==1.0.0
urllib3==1.26.3
